remove logic from controller to service in cashmachine
